Cleanup the code and adding token role check functionality.

+++ app/Http/Controllers/LoanController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 
 class LoanController extends Controller
 {
+    const UNAUTHORIZED_ACTION_MESSAGE = 'User unauthorized to perform this action.';
     /**
      *
      */
@@ -25,16 +26,22 @@ class LoanController extends Controller
             'amount' => 'required|numeric|min:1',
             'interest_rate' => 'required|numeric|min:0',
             'duration' => 'required|integer|min:1',
+            'borrower_user_id' => 'required|numeric|exists:users,id'
         ]);
+        $user = auth()->user();
+        if ($user->tokenCan('abilities:loan-create,loan-update,loan-delete')) {
+            $loan = Loan::create([
+                'user_id' => $user->id,
+                'amount' => $request->amount,
+                'interest_rate' => $request->interest_rate,
+                'duration' => $request->duration,
+                'borrower_user_id' => $request->borrower_user_id
+            ]);
+            return response()->json($loan, 201);
+        } else {
+            return response()->json(['message' => self::UNAUTHORIZED_ACTION_MESSAGE], 400);
+        }
 
-        $loan = Loan::create([
-            'user_id' => auth()->user()->id,  // Assuming you're associating the loan with the authenticated user
-            'amount' => $request->amount,
-            'interest_rate' => $request->interest_rate,
-            'duration' => $request->duration,
-        ]);
-
-        return response()->json($loan, 201);
     }
 
     /**
@@ -68,7 +75,8 @@ class LoanController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $loan = auth()->user()->loans()->find($id);
+        $user = auth()->user();
+        $loan = $user->loans()->find($id);
 
         if (!$loan) {
             return response()->json(['message' => 'Loan not found'], 404);
@@ -79,14 +87,18 @@ class LoanController extends Controller
             'interest_rate' => 'required|numeric|min:0',
             'duration' => 'required|integer|min:1',
         ]);
+        if ($user->tokenCan('abilities:loan-create,loan-update,loan-delete')) {
+            $loan->update([
+                'amount' => $request->amount,
+                'interest_rate' => $request->interest_rate,
+                'duration' => $request->duration,
+            ]);
+            return response()->json($loan);
+        } else {
+            return response()->json(['message' => self::UNAUTHORIZED_ACTION_MESSAGE], 400);
+        }
 
-        $loan->update([
-            'amount' => $request->amount,
-            'interest_rate' => $request->interest_rate,
-            'duration' => $request->duration,
-        ]);
 
-        return response()->json($loan);
     }
 
     /**
@@ -95,13 +107,18 @@ class LoanController extends Controller
      */
     public function destroy($id)
     {
-        $loan = auth()->user()->loans()->find($id);
+        $user = auth()->user();
+        $loan = $user->loans()->find($id);
 
         if (!$loan) {
             return response()->json(['message' => 'Loan not found'], 404);
         }
+        if ($user->tokenCan(self::UNAUTHORIZED_ACTION_MESSAGE)) {
+            $loan->delete();
+            return response()->json(['message' => 'Loan deleted successfully']);
+        } else {
+            return response()->json(['message' => self::UNAUTHORIZED_ACTION_MESSAGE], 400);
+        }
 
-        $loan->delete();
-        return response()->json(['message' => 'Loan deleted successfully']);
     }
 }
